replace (unsigned long) with (uintptr_t)

+++ src/packet/adva_tlv_header.c
@@ -39,7 +39,7 @@ adva_tlv_header_new(void)
 {
     adva_tlv_header_t *header = (adva_tlv_header_t *) header_storage_new(&storage);
 
-    LOG_PRINTLN(LOG_HEADER_NTP, LOG_DEBUG, ("NTP header new 0x%016" PRIxPTR, (unsigned long) header));
+    LOG_PRINTLN(LOG_HEADER_NTP, LOG_DEBUG, ("NTP header new 0x%016" PRIxPTR, (uintptr_t) header));
 
     return header;
 }
@@ -49,7 +49,7 @@ adva_tlv_header_free(header_t *header)
 {
     if (header->next != NULL)   header->next->klass->free(header->next);
 
-    LOG_PRINTLN(LOG_HEADER_NTP, LOG_DEBUG, ("NTP header free 0x%016" PRIxPTR, (unsigned long) header));
+    LOG_PRINTLN(LOG_HEADER_NTP, LOG_DEBUG, ("NTP header free 0x%016" PRIxPTR, (uintptr_t) header));
 
     header_storage_free(header);
 }
