+++ src/packet/icmpv4_header.c
@@ -43,7 +43,7 @@ icmpv4_header_new(void)
 {
     icmpv4_header_t *header = (icmpv4_header_t *) header_storage_new(&storage);
 
-    LOG_PRINTLN(LOG_HEADER_IPV4, LOG_DEBUG, ("ICMPv4 header new 0x%016" PRIxPTR, (unsigned long) header));
+    LOG_PRINTLN(LOG_HEADER_IPV4, LOG_DEBUG, ("ICMPv4 header new 0x%016" PRIxPTR, (uintptr_t) header));
 
     return header;
 }
@@ -53,7 +53,7 @@ icmpv4_header_free(header_t *header)
 {
     if (header->next != NULL)   header->next->klass->free(header->next);
 
-    LOG_PRINTLN(LOG_HEADER_IPV4, LOG_DEBUG, ("ICMPv4 header free 0x%016" PRIxPTR, (unsigned long) header));
+    LOG_PRINTLN(LOG_HEADER_IPV4, LOG_DEBUG, ("ICMPv4 header free 0x%016" PRIxPTR, (uintptr_t) header));
 
     header_storage_free(header);
 }
