+++ src/packet/udpv4_header.c
@@ -40,7 +40,7 @@ udpv4_header_new(void)
 {
     udpv4_header_t *header = (udpv4_header_t *) header_storage_new(&storage);
     
-    LOG_PRINTLN(LOG_HEADER_UDPV4, LOG_DEBUG, ("UDPv4 header new 0x%016" PRIxPTR, (unsigned long) header));
+    LOG_PRINTLN(LOG_HEADER_UDPV4, LOG_DEBUG, ("UDPv4 header new 0x%016" PRIxPTR, (uintptr_t) header));
     
     return header;
 }
@@ -50,7 +50,7 @@ udpv4_header_free(header_t *header)
 {
     if (header->next != NULL)   header->next->klass->free(header->next);
     
-    LOG_PRINTLN(LOG_HEADER_UDPV4, LOG_DEBUG, ("UDPv4 header free 0x%016" PRIxPTR, (unsigned long) header));
+    LOG_PRINTLN(LOG_HEADER_UDPV4, LOG_DEBUG, ("UDPv4 header free 0x%016" PRIxPTR, (uintptr_t) header));
     
     header_storage_free(header);
 }
